% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{approxm}
\alias{approxm}
\title{Interpolate multiple columns of a \code{data.frame}}
\usage{
approxm(data, n, method = "linear")
}
\arguments{
\item{n}{the number of points to interpolate}

\item{method}{the method to interpolate by, \code{"linear"} or 
\code{"spline"}}

\item{data1}{a \code{data.frame} with columns for interpolation}
}
\value{
Returns a \code{data.frame} of all \code{data1} variables
  interpolated to \code{n} rows.
}
\description{
Interpolate multiple columns of a \code{data.frame}
}
\seealso{
\code{\link{ks.test}} is the underlying function and has more
  options
}

