% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{cor.testm}
\alias{cor.testm}
\title{Correlation test for multiple variables - of the same name - in separate 
data.frames}
\usage{
cor.testm(data1, data2, method)
}
\arguments{
\item{data1}{a \code{data.frame} with N variables}

\item{data2}{a \code{data.frame} with the same N variables as \code{data1}}

\item{method}{a correlation method, either \code{method = "pearson"} or 
\code{"spearman"}}
}
\value{
Returns a \code{data.frame} of correlation test data for N variables
}
\description{
Correlation test for multiple variables - of the same name - in separate 
data.frames
}
\seealso{
\code{\link{cor.test}} is the underlying function and has more
  options
}

