% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generate_behaviors}
\alias{generate_behaviors}
\title{Generate Random Binary Vectors}
\usage{
generate_behaviors(nrow, nBehaviors, prStart = 0.1, prEnd = 0.3,
  prBy = 0.001)
}
\arguments{
\item{nrow}{bumber of rows to generate}

\item{nBehaviors}{number of columns to generate}

\item{prStart}{starting probability}

\item{prEnd}{ending probability}

\item{prBy}{probability to sequence by}
}
\value{
a matrix of binary values
}
\description{
Generate Random Binary Vectors
}
\examples{
generate_behaviors(10, 10)
}

