% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generate_dates}
\alias{generate_dates}
\title{Generate a sequence of dates}
\usage{
generate_dates(startDate, nDays, by = "day")
}
\arguments{
\item{startDate}{a character string in the format "YYYY-MM-DD"}

\item{nDays}{length of output vector}

\item{by}{days or weeks}
}
\value{
Sequence of dates starting at startDate with length nDays by daily or
  weekly increments
}
\description{
Generate a sequence of dates
}
\examples{
generate_dates("2016-06-05", 31, by = "day")

}

