% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{import.data}
\alias{import.data}
\title{Import Data Pop Up}
\usage{
import.data(filename = file.choose())
}
\arguments{
\item{filename}{a string indicating the path of the file containing the data.
By default, will open a pop-up so that the user can choose a file with the
GUI.}
}
\value{
A data.frame ready to be converted into freqprof class (see function
  \code{freqprof}).
}
\description{
This function reads a file, whose extension is either csv, bin or fpw, and
imports it as a data.frame.
}
\examples{
\donttest{
# Select a file
import.data()

# Open bin file included in package extdata folder
filepath <- system.file("extdata", "S58-1-1.bin", package = "FreqProf")
import.data(filepath)
}
}

