% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{ks.testm}
\alias{ks.testm}
\title{Kolmogorov-Smirnov test for multiple variables - of the same name - in 
separate data.frames}
\usage{
ks.testm(data1, data2, vars)
}
\arguments{
\item{data1}{a \code{data.frame} with N variables}

\item{data2}{a \code{data.frame} with the same N variables as data1}

\item{vars}{a \code{vector} of the N variable names}
}
\value{
Returns a \code{data.frame} with Kolmogorov-Smivnov test data for N
  variables
}
\description{
Kolmogorov-Smirnov test for multiple variables - of the same name - in 
separate data.frames
}
\seealso{
\code{\link{ks.test}} is the underlying function and has more
  options.
}

