% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeData.R
\name{movfun}
\alias{movfun}
\title{Internal function for Generating Moving Sum or Proportion}
\usage{
movfun(x, n, s, r, fun)
}
\arguments{
\item{x}{data}

\item{n}{window length}

\item{s}{step}

\item{r}{resolution}

\item{fun}{"sum" or "proportion"}
}
\value{
Returns a list containing the processed data into $movfun, and the 
  associated panels into $panels. Passes list to \code{\link{freqprof}}.
}
\description{
Internal function in \code{\link{freqprof}} that is used to generate moving 
sum or proportion data.
}
\examples{
movfun(as.numeric(runif(10)>.5),2,1,1,"sum")
}

