% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeData.R
\name{print.freqprof}
\alias{print.freqprof}
\title{Prints the data contained into a variable of class freqprof}
\usage{
\method{print}{freqprof}(data.freqprof)
}
\arguments{
\item{data.freqprof}{the data}
}
\value{
NULL.
}
\description{
Prints the data contained into a variable of class freqprof
}
\examples{
print.freqprof(freqprof(import.data()))
}

