% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeData.R
\name{radj}
\alias{radj}
\title{Internal function for Resolution Adjustment}
\usage{
radj(x, r)
}
\arguments{
\item{x}{data data passed from \code{\link{freqprof}}}

\item{r}{resolution passed from \code{\link{freqprof}}}
}
\value{
Resolution adjustment.
}
\description{
Internal function in \code{\link{freqprof}} that is used to modify data resolution.
}
\examples{
radj(as.numeric(runif(10)>.5),2)
}

