% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{read.fpw}
\alias{read.fpw}
\title{Reads the data in the file "filename", which is supposed to be a .fpw file}
\usage{
read.fpw(filename)
}
\arguments{
\item{filename}{a string indicating the path of the file.}
}
\value{
A data.frame of the .fpw data.
}
\description{
Reads the data in the file "filename", which is supposed to be a .fpw file
}

