% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{s58}
\alias{s58}
\title{Occurrence/nonoccurrence data for four behaviors from a single subject}
\format{A data.frame with 3092 rows and 4 variables 
\describe{ 
  \item{V1}{Behavior 1}
  \item{V2}{Behavior 2}
  \item{V3}{Behavior 3}
  \item{V4}{Behavior 4}
}}
\usage{
s58
}
\description{
A dataset containing the occurrence/nonoccurence data for four behaviors of a
single subject over 3092 observations. Each behavior corresponds to touches
on areas of a touchpad.
}
\keyword{datasets}

