% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeData.R
\name{tail.freqprof}
\alias{tail.freqprof}
\title{Prints the tail of the data contained into a variable of class freqprof}
\usage{
\method{tail}{freqprof}(data.freqprof)
}
\arguments{
\item{data.freqprof}{the data}
}
\value{
NULL.
}
\description{
Prints the tail of the data contained into a variable of class freqprof
}
\examples{
tail.freqprof(freqprof(import.data()))
}

