% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotData.R
\name{ggplot_fp}
\alias{ggplot_fp}
\title{Internal ggplot Wrapper to Graph Frequency Profiles}
\usage{
ggplot_fp(data1, resolution = resolution, step = step, yAxis = yAxis,
  xAxisUnits = xAxisUnits, xmin = xmin, xmax = xmax,
  tick.every = tick.every, label.every = label.every)
}
\arguments{
\item{data1}{}

\item{resolution}{}

\item{step}{}

\item{yAxis}{}

\item{xAxisUnits}{}

\item{xmin}{}

\item{xmax}{}

\item{tick.every}{}

\item{label.every}{}
}
\value{

}
\description{
Internal ggplot Wrapper to Graph Frequency Profiles
}

