% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeData.R
\name{movfun}
\alias{movfun}
\title{Moving function}
\usage{
movfun(x, n, s, r, fun)
}
\arguments{
\item{x}{data}

\item{n}{window length}

\item{s}{step}

\item{r}{resolution}

\item{fun}{"sum" or "proportion"}
}
\value{
Used for computing moving function. Returns a list containing the processed data into $movfun, and the associated panels into $panels.
}
\description{
Moving function
}
\examples{
movfun(as.numeric(runif(10)>.5),2,1,1,"sum")
}

