% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeData.R
\name{movfun}
\alias{movfun}
\title{Internal function for Generating Moving Sum or Proportion}
\usage{
movfun(x, n, s, r, fun)
}
\arguments{
\item{x}{data passed from \code{\link{freqprof}}}

\item{n}{window length passed from \code{\link{freqprof}}}

\item{s}{step size passed from \code{\link{freqprof}}}

\item{r}{resolution passed from \code{\link{freqprof}}}

\item{fun}{"sum" or "proportion" passed from \code{\link{freqprof}}}
}
\value{
Returns a list containing the processed data into $movfun, and the 
  associated panels into $panels. Passes list to \code{\link{freqprof}}.
}
\description{
Internal function in \code{\link{freqprof}} that is used to generate moving 
sum or proportion data.
}

