% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotData.R
\name{plot_freqprof}
\alias{plot_freqprof}
\title{Plot Frequency Profiles.}
\usage{
plot_freqprof(data.freqprof, yAxis = NULL, xAxisUnits = "sec",
  panel.in = T, panel.out = T, gg = F, multiPlot = F,
  tick.every = round(length(data.freqprof$data$time) / 31), label.every = 3)
}
\arguments{
\item{data.freqprof}{data formated into \code{freqprof} class.}

\item{yAxis}{a string giving the legend of the y-axis.}

\item{xAxisUnits}{a string indicating which unit has been used. By default, 
"sec".}

\item{panel.in}{a boolean indicating if the first panel has to be plotted.}

\item{panel.out}{a boolean indicating if the third panel has to be plotted}

\item{gg}{if TRUE, will use the 'ggplot2' package. By default, gg = FALSE.}

\item{multiPlot}{if TRUE, will plot each behavior in a distinct panel. By 
default, multiPlot = FALSE.}

\item{tick.every}{the spacing between each tick. By default, N/30 where N is 
the number of time units.}

\item{label.every}{label every X ticks, where X = label.every. By default, 
label.every = 3.}
}
\value{
The function plots the data but does not return anything.
}
\description{
Plot data that has already been converted to class \code{freqprof} with 
\code{\link{freqprof}}.
}
\examples{
data(s58)
plot_freqprof(freqprof(s58))
}

