% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeData.R
\name{radj}
\alias{radj}
\title{Internal function.}
\usage{
radj(x, r)
}
\arguments{
\item{x}{data}

\item{r}{resolution}
}
\value{
Helpful for \code{movfun}.
}
\description{
Internal function.
}
\examples{
radj(as.numeric(runif(10)>.5),2)
}

