% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FreqProf_sourcecode.R
\name{radj}
\alias{radj}
\title{Internal function.}
\usage{
radj(x, r)
}
\arguments{
\item{x}{}

\item{r}{}
}
\value{
Helpful for \code{movfun}.
}
\description{
Internal function.
}
\examples{
radj(as.numeric(runif(10)>.5),2)
}

